+++ twm_window_struct.h
@@ -104,8 +104,34 @@ struct TwmWindow {
 
 	/// @}
 
-	char *name;       ///< Window name.  From WM_NAME property.
-	char *icon_name;  ///< Icon name.  From WM_ICON_NAME property.
+	char *name;       ///< Current window name.  Points into TwmWindow::names.
+	char *icon_name;  ///< Current icon name. Points into TwmWindow::names.
+
+	/// Various sources of window/icon names.  These are the values from
+	/// the various window properties we look at to get the results.  The
+	/// TwmWindow::name and TwmWindow::icon_name point to the currently
+	/// active element in here.
+	struct _names {
+		char *ctwm_wm_name; ///< Name from override CTWM_WM_NAME property
+#ifdef EWMH
+		char *net_wm_name;  ///< Name from EWMH _NET_WM_NAME property
+#endif
+		char *wm_name;      ///< Name from ICCCM WM_NAME property
+
+		/// Icon name from override CTWM_WM_ICON_NAME property
+		char *ctwm_wm_icon_name;
+#ifdef EWMH
+		/// Icon name from EWMH _NET_WM_ICON_NAME property
+		char *net_wm_icon_name;
+#endif
+		char *wm_icon_name; ///< Icon name from WM_ICON_NAME property
+
+		/// Whether an icon name property has been set.  Since we default
+		/// the icon name to the window name when nothing is given, this
+		/// flag allows the window-name-setting code to know when it
+		/// needs to re-kick the icon-name-setting.
+		bool icon_set;
+	} names; ///< \copydoc TwmWindow::_names
 
 	/// \addtogroup win_frame Window frame bits
 	/// @{