+++ win_utils.h
@@ -27,6 +27,10 @@ void ConstrainByBorders(TwmWindow *twmwin, int *left, int width,
 void WarpToWindow(TwmWindow *t, bool must_raise);
 void send_clientmessage(Window w, Atom a, Time timestamp);
 XWMHints *gen_synthetic_wmhints(TwmWindow *win);
+bool set_window_name(TwmWindow *win);
+void apply_window_name(TwmWindow *win);
+bool set_window_icon_name(TwmWindow *win);
+void apply_window_icon_name(TwmWindow *win);
 
 
 #endif /* _CTWM_WIN_UTILS_H */